Moved AssassinHomePage to its own dart file
